Replace SQLAlchemy Numeric Type with Real
SQLAlchemy was using a custom Numeric Type for Numeric Values,
and was causing warnings, as Numerics could on some platforms store
Decimals, not just floats.

Using Real instead confirms that a float64 is saved.

+++ nqm/iotdatabase/_sqlitealchemyconverter.py
@@ -19,7 +19,7 @@ sqlalchemyMap: t.Dict[
 ] = {
     _sqliteconstants.SQLITE_TYPE.INTEGER: sqlalchemy.types.Integer,
     _sqliteconstants.SQLITE_TYPE.TEXT: sqlalchemy.types.String,
-    _sqliteconstants.SQLITE_TYPE.NUMERIC: sqlalchemy.types.Numeric,
+    _sqliteconstants.SQLITE_TYPE.NUMERIC: sqlalchemy.types.REAL,
     _sqliteconstants.SQLITE_TYPE.REAL: sqlalchemy.types.REAL,
     _sqliteconstants.SQLITE_TYPE.INTEGER: sqlalchemy.types.Integer
 }
